Update queries Moved single page queries to components. relates to #3
