+++ app/(site)/page.tsx
@@ -4,7 +4,7 @@ import { load_Settings } from '@/sanity/loader/loader';
 const Page = async() => {
 	const settings = await load_Settings()
 
-	if (!settings || !settings.homepage) {
+	if (!settings || !settings.homepage?.pathname?.current) {
 		return (
 			<div className="text-center">
 				No homepage set.
@@ -12,7 +12,7 @@ const Page = async() => {
 		)
 	}
 
-	return <Pages.PageRoute params={{ slug: [settings.homepage.slug] }} />
+	return <Pages.PageRoute params={{ slug: [settings.homepage.pathname?.current] }} />
 }
 
 export default Page
