+++ components/Pages/PagePreview.tsx
@@ -1,11 +1,17 @@
 'use client'
 
 import { Page } from '@/components/Pages/Page'
-import { useSingle_Page } from '@/sanity/loader/useQuery'
+import { useQuery } from '@/sanity/loader/useQuery'
+import { pageQuery } from '@/sanity/queries/queries'
 import { type QueryResponseInitial } from '@sanity/react-loader'
 
 type Props = {
 	initial: QueryResponseInitial<PagePayload | null>
+}
+
+export function useSingle_Page(initial: QueryResponseInitial<PagePayload | null>) {
+	const pathname = `${initial?.data?.pathname?.current}`
+	return useQuery<PagePayload | null>(pageQuery, { pathname }, { initial })
 }
 
 const PagePreview = (props: Props) => {
