+++ sanity/queries/queries.ts
@@ -8,7 +8,7 @@ export const settingsQuery: string = groq`
 			${partial_ImageObject}
 		},
 		homepage->{
-			"slug":slug.current,
+			pathname
 		},
 		navigation[]{
 			// ...,
@@ -19,6 +19,13 @@ export const settingsQuery: string = groq`
 				${partial_Sections}
 			}
 		}
+	}
+`
+
+export const pageQuery: string = groq`
+	*[_type=='page' && pathname.current == $pathname][0] {
+  	...,
+		${partial_Sections}
 	}
 `
 
@@ -33,15 +40,6 @@ export const staticPagesQuery: string = groq`
 export const staticArchivesQuery: string = groq`
 	*[_type == 'archive'] {
 		...,
-		${partial_Sections}
-	}
-`
-
-export const pageQuery: string = groq`
-	*[_type=='page' && pathname.current == $pathname][0] {
-  	...,
-		// "slug":slug.current,
-
 		${partial_Sections}
 	}
 `
